Remove duplicated files with tests
